requirements.txt
# Main application dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
supabase==2.0.0
python-dotenv==1.0.0
markdown==3.5.1

# Database dependencies
psycopg2-binary==2.9.7
sqlalchemy==2.0.23

# Machine Learning dependencies
tensorflow>=2.16.0,<2.20.0
keras-tuner>=1.4.7
pandas>=2.1.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# API and data processing
requests==2.31.0
tqdm==4.66.1
pytz==2023.3

# AI/LLM dependencies
google-generativeai==0.3.0
langchain==0.0.340
langchain-community==0.0.1
openai==1.3.6
tabulate==0.9.0

# Firebase (optional)
firebase-admin==6.2.0
